requirements.txt
django
djangorestframework
django-cors-readers